photon-lib/py/build_requirements.txt
nanobind~=2.1.0
typing-extensions
wpilib~=2024.3.2.1
robotpy-wpimath~=2024.3.2.1
robotpy-apriltag~=2024.3.2.1
pyntcore~=2024.3.2.1
build
setuptools
wheel
pytest
